APP-261 rename misused .d.ts files to .ts in apps/webapp/src
22 of 23 hand-authored .d.ts files renamed to .ts so TypeScript stops
silently ignoring their contents under skipLibCheck. vite-env.d.ts is
left alone (legitimate ambient declarations). widgetState.d.ts is
deferred — renaming it surfaces ~130 type errors across WidgetState
consumer sites (missing StakeAction/StUSDS/MorphoVault union members,
null-reset pattern, downstream narrowing) that warrant a separate audit
per the issue's "split if scope grows" clause.

Latent bugs surfaced and fixed in this PR:
- hooks/hooks.ts: UseSendBatchTransactionFlowParameters chainId had
  `| undefined = undefined`, which violates wagmi's SendCallsParameters
  constraint. Dropped the | undefined.
- hooks/upgrade/upgrade.ts: UpgradeResponses was missing the required
  generic on UpgradeResponse. Made UpgradeResponses generic; updated
  useUpgradeHistory.ts to use per-key types.
- modules/app/types/Widgets.ts: JSX.Element → React.JSX.Element;
  imported missing IconProps.
- widgets/config/types/widgets-config.ts: removed two unused, syntactically
  invalid type aliases (referenced undefined Token/TokenForChain with
  invalid indexer syntax). Added TokenForChain import.
- widgets/config/default-config.ts: balancesTokenList ETH entries were
  passing the raw Token (nested per-chain address map) where
  TokenForChain[] is expected. Mirrored the tradeTokenList pattern.
