+++ apps/webapp/src/hooks/morpho/index.ts
@@ -23,7 +23,7 @@ export {
   type MorphoIdleLiquidityAllocation,
   type MorphoVaultAllocationsData,
   type MorphoVaultAllocationsHook
-} from './morpho.d';
+} from './morpho';
 export {
   useMorphoVaultRewards,
   type MorphoVaultReward,
