+++ apps/webapp/src/hooks/stusds/index.ts
@@ -33,7 +33,7 @@ export type { StUsdsHistoryHook } from './useStUsdsHistory';
 export { useStUsdsChartInfo } from './useStUsdsChartInfo';
 
 // Types
-export type { StUsdsHistoryItem, StUsdsVaultMetrics, StUsdsUserMetrics } from './stusds.d';
+export type { StUsdsHistoryItem, StUsdsVaultMetrics, StUsdsUserMetrics } from './stusds';
 
 // Provider Abstraction Layer (Curve pool integration)
 export * from './providers';
