+++ apps/webapp/src/widgets/TradeWidget/lib/types.ts
