+++ apps/webapp/src/widgets/shared/types/txCardCopyText.ts
